requirement.txt
accelerate==0.24.1
Flask==3.0.0
numpy==1.26.1
Pillow==10.1.0
torch==2.1.0
transformers==4.35.0
